Refactored query package up to request package
