+++ messaging-service/internal/application/request/chat.go
@@ -1,26 +1,26 @@
-package query
+package request
 
 import "github.com/google/uuid"
 
-type CreateGroupRequest struct {
+type CreateGroup struct {
 	Admin   uuid.UUID
 	Members []uuid.UUID
 	Name    string
 }
 
-type UpdateGroupInfoRequest struct {
+type UpdateGroupInfo struct {
 	ChatID      uuid.UUID
 	Name        string
 	Description string
 }
 
-type CreateSecretGroupRequest struct {
+type CreateSecretGroup struct {
 	Admin   uuid.UUID
 	Members []uuid.UUID
 	Name    string
 }
 
-type UpdateSecretGroupInfoRequest struct {
+type UpdateSecretGroupInfo struct {
 	ChatID      uuid.UUID
 	Name        string
 	Description string
