+++ messaging-service/internal/application/request/update.go
@@ -1,0 +1,4 @@
+package request
+
+type SendTextMessage struct {
+}
